Update scripts to support astex datasets and chai boltz models

+++ scripts/run_alphafold3/run_alphafold3_part_1.sh
@@ -2,10 +2,19 @@
 
 PART_ID=1
 
+# Check if model type is provided as argument
+if [ $# -eq 0 ]; then
+    echo "Error: Please provide dataset as argument"
+    echo "Usage: $0 <dataset>"
+    exit 1
+fi
+
+DATASET="$1"
+
 # Run AlphaFold3
 docker run -it --rm --gpus all --shm-size=32g -e CUDA_VISIBLE_DEVICES=${PART_ID} \
-	-v /home/jiangyize/workspace/projects/protein_ligand_docking_benchmark/data/microcyto/posebusters/alphafold3/input/part_${PART_ID}:/root/af_input \
-	-v /home/jiangyize/workspace/projects/protein_ligand_docking_benchmark/data/microcyto/posebusters/alphafold3/output:/root/af_output \
+	-v ./data/microcyto/${DATASET}/alphafold3/input/part_${PART_ID}:/root/af_input \
+	-v ./data/microcyto/${DATASET}/alphafold3/output:/root/af_output \
 	-v /data/dataset/alphafold3/models:/root/models \
 	-v /data/dataset/alphafold3/databases:/root/public_databases \
 	brandonsoubasis/alphafold3 \
